Add --waitforpid option to test framework.
Sometimes we want to do something with the running bitcoind process
after it has started but before the main test runs; for example attach
gdb to the running process and set some breakpoints.

This change introduces a --waitforpid command line option to the
functional tests that spin up bitcoind processes which causes that
test to print the pid of the launched process and then pause before
continuing.