Fix race in ActivateBestChain()
If ActivateBestChain is called by multiple threads in parallel it
can result in attempts made to set the latest tip to different
candidates with most work.

Add a check to see if someone else has updated the tip from under
us and if so recalculate the best work target tip we are trying
to update to.

Note: I don't actually think this race is possible to trigger
at present since all paths to here are currently executed in
a single thread. But as we add parallelism to the flow it
certainly is possible to be triggered.