Make block validity checking from new mining candidate optional.
The block validity checking on the newly created mining candidate
is pointless and expensive. The new block is made from transactions
we have already validated, and all the checks we perform here
are duplicated at the point when a solution is checked, which
makes much more sense.

Make whether we run TestBlockValidity() for a new block configurable.