Fix crash with new mining API.
The new mining API has a strange 2 stage construction process for
the mining candidates: 1) Create the candidate, 2) Set the block
it is for. This leaves the possibility of a period of time when
a mining candidate doesn't actually have a valid block set
within it. An attempt to use the block reference during this
window would result in a failure.

Change it so that the candidate is only created once we have
the block it is for.