Record block header fields with mining candidate.
Some block header fields need to be updated on submission of a
mining solution, but those fields are currently shared amongst
several candidates.

Pull out updatable fields and store them in the candidate so
that if they have been updated by some other miner or a later
candidate we still remember our values.