Fix possible deadlock identified for Clang thread sanitiser.
There's a theoretical deadlock in the journaling code identified by
Clang, although it's very unlikely to trigger. Fix it by releasing
the old journal read lock while we update to a new journal and
then re-taking it.

Also just move some logging outside of the scope of the journal lock
while we're at it.