Fix race in miner_tests at shutdown.
The journaling block assembler used by the miner_tests might not be
shutdown before the block index it relies on was destroyed by the
unit test framework.

This change gets rid of the global singleton journaling block assembler
and makes it owned by a mining factory class that has a well defined
and controllable lifetime.