Late configuration for time locked mempool.
the time locked pool is owned by the standard mempool which is
a global static object with undefined construction order. It tries
to configure itself from the gArgs object which is another
global static which might not have been initialised first.

Add a method loadconfig() to allow us to control when the config
is read and cached.