Remove racy block stalling test.
On a suffuciently fast modern machine sending even a very large
block over the loopback interface is not slow enough to trigger a
stall with the new changes to prevent stalling. Remove the test
that tries to trigger a stall since it never really added much
to the test suite and it will be forever racy.