Perform early clearing out of mapAskFor.
CNode::mapAskFor is a priority queue containing items we should ask that
peer for, ordered on the earliest time we should ask for them. The time
gap between duplicate requests for the same item to different peers is
2 minutes.

Items are removed from mapAskFor only when the time to request that item
expires. This means that if we receive INVs for the same item from different
peers, then mapAskFor for one of those peers will contain entries for items
we have already asked for for 2 minutes.

If the P2P is busy with many transactions we will get duplicate INVs for
the same transaction from many peers, and mapAskFor can quickly fill up
with entries that won't expire for 2 minutes or more.

This commit make a small improvement to SendGetDataNonBlocks so that we
look ahead a little when scanning mapAskFor for items to request, and
will expire early any entries we find immediately following items it
is time to request if we have already seen those items from elsewhere.

We don't completely rescan all items in mapAskFor every time because
that could be quite expensive.