Rename BLOCK and TRANSACTION streams.
Rename the BLOCK and TRANSACTION streams to DATA1 and DATA2. The old
names were too specific, the new names are more general and better
reflect that they are actually just 2 data channels that can be used
for whatever purpose a particular policy deems fit.

Also add a further couple of DATA3 and DATA4 stream types so that
they are already there and available if a new stream policy is implemented
that wants to use them.