Increase the default maximum send/receive buffer sizes.
The old limits were from the small block world and limited the buffer
sizes to 1Mb and 5Mb respectively which would cause the node to
start throttling back as soon as any sizeable block were queued.

Increase them to 500Mb to better reflect the sort of block sizes
we are regularly targetting.