Handle exception in ActivateBestChainStep.
If block validation is cancelled (for example during shutdown of the
node) an exception is thrown from within ConnectBlock. That exception
causes the disconnectpool from inside ActivateBestChainStep to be
destroyed while non-empty and an assert in its destructor to trigger.

Handle the exception gracefully inside ActivateBestChainStep and
rethrow it so that it can also be handled at the expected site.