Prevent blocking the socket handling thread.
The socket handling thread is generally free from needing to take cs_main so
should be capable of running all the time even when other threads might be
blocked. The only exception to this is the case when it tries to tidy up old
disconnected peers.

Modify the DeleteNode -> FinalizeNode code path so that if we are unable to
take cs_main we don't block waiting, but instead just carry on and try again
in a short while.