Reduce the CTxMemPool public API by moving ancestor/descendant calculation to private methods.
This change temporarily makes transaction validation more expensive. Because the
set of a transactions's ancestors is no longer exposed to the validator, it will
be recalculated when the transaction is committed to the mempool. However, the
final solution for CORE-130 will introduce different criteria for ancestor and
descendant limits that will mitigate this performance degradation.