Check for being paused for sending and receiving.
If we are paused for both sending and receiving the message processing
can grind to a halt because we won't take any messages off the receiving
queue until the other peer begins to empty their socket buffer. This
can become a catch-22 if both sides are waiting for the other to move
first.

If we're paused for both sending and receiving then allow a temporary
increase in the maximum allowed send buffer size so that we can go on
and process some more incoming messages and break the deadlock.