bsv-ptv-p2p: drop racy assert
Since the reject list length can not be negative the assert can be simply removed.

[explanation by Arkadiusz, why the check can be relaxed]

At this stage, during test's execution we simply check an expected number of
rejected messages cased by 'txn-double-spend-detected' reject reason (as a prof
that double spend has happened but not a mempool conflict - which is easier to
get). When, I was testing it on my environment, I was constantly getting > 0
number of such rejects, that's why there is a condition > 0.

What is more, in this test case (TC5), this assert is the last and final check.
Before that check we verify if:

- all txs from the nonstd_txs result set are accepted
  `self.check_mempool_with_subset(conn.rpc, nonstd_txs, timeout=30)`

- there are tc5_num_of_subsets more transaction in the mempool (random txns from
  the ds_txs set)

  `assert_equal(conn.rpc.getmempoolinfo()['size'], len(nonstd_txs)+tc5_num_of_subsets)`

- only tc5_num_of_subsets txns are allowed to be in the mempool from the given
  ds set.

  `assert_equal(len(self.check_intersec_with_mempool(conn.rpc, ds_txs)), tc5_num_of_subsets)`

The checks above simply guarantee that there is an expected number of
transactions in the mempool (including a single txn from a given double spend
set - 10 in this test case).

As you are not constantly getting 'txn-double-spend-detected' reject messages
(as in your executions 25% of the time all transactions are detected as
'txn-mempool-conflict' for which we do not create a reject messages), due to
runtime-dependant conditions, I suggest that you to change the condition to >= 0.