Removal of swapping backends for coin views
Remove CCoinsViewCache:SetBackend() to prevent swapping of
underlying views. We need to keep underlying views connected
during the while life time of CCOinViewCache in order to
delay load the coin scripts in later commits.
Legacy code swapped backend after caching was performed so that
the locks did not have to be held for the entire duration of
the processing task so this commit is a pesimization which will
be removed in later commit by changing the code so that the
coins pool lock will be held only when needed to access coin
for the first time.

NOTE: Change to TxnValidation relies on view not accessing coins that
were not cached while holding pool.smtx lock which is a dangerous
assumption - but valid for current implementation.