Split of multi and single threaded coins views
Main motivation for the split is that multithreaded version
is only used for pcoinsTip and requires a different API than
the single threaded version - some functions that are not
needed by pcoinsTip are also hard to implement correctly in
combination with hierarchy locking mechanism that is part of
a future commit. Secondary reason was different cache usage
patterns (will require less coin copying in a future commit)
and performance (besides memory copying there is also no need
for locks and other synchronization mechanisms in the single
threaded version).

Commit also removes two unrealistic tests as single threaded
providers are never - and from this commit can't be - chained
in bitcoind. Besides that a future commit will introduce
implicit provider/view hierarchical locking which will make
the two tests even less realistic as only one branch and even
there only the top provider will be able to make changes to
the hierarchy guaranteed by the api and implicit locks.