Revert "Remove SaltedOutpointHasher"
This reverts commit d8829b7082c508a29e6a00e005d30350611f315c.

In CORE-1022 during testing it was noticed that there is a
negligible performance change in the profiler for find function
calls on the coins store when using std::map instead of
std::unordered_map. After further consideration this was due
to the fact that there weren't many coins in the store which
changes if there are coins from multiple blocks not flushed to
disk so it was decided to revert the commit.