"-disablebip30checks" command line param
These checks are performed on the ConnectBlock. On the main-net they are performed only on the several blocks between enabling of BIP30 and BIP34. On the other networks (stn, testnet and regtest) it is performed always. This brings huge performance penalty when connecting a block and does not reflect the real node capability. This flag turns checks off.