Fixup of the CPFP Group ID
Group id was the same as the paying tx id. This caused problems when we were removing txs from the mempool as we first removed paying tx. After that when deleting its ancestors we asked for group id (paying tx id) which tried to access txid if the removed (paying) tx.