Fix shutdown ordering race.
Make it explicit in the Shutdown() function that the block index is freed
after the block assembler has shutdown. This is already the case, but relies
on the global object from validation.cpp cleaning up in the destructor, which
might not be the case if (for example) tests are being run from
bitcoin_fuzzing.

While we're here, also explicitly shutdown the merkle tree store for added
safety.