Add a leaky bucket algorithm class.
A simple implementation of the leaky bucket algorithm.

Values are added to the total stored in the bucket and if that total exceeds
a defined threshold the bucket overflows. Meanwhile over time the bucket
leaks and the total steadily drains out.

This can be used to measure if the frequency of some event exceeds a set
limit.