Fix: we should detach a CoinsDB view before re-locking cs_main
In CORE-1165 unlocked region inside ConnectBlock was expanded
and this causes that in some cases during parallel block
validation we try to re-obtain cs_main lock while still holding
CoinsDB lock which can cause dead locks due to locking order.
This commit also moves CTemporaryLeaveCriticalSectionGuard class
to the only location where it is used as it is dangerous to use
to begin with and will hopefully be obsolete sooner rather than
later.