Fix: mark CBlockIndex instance as dirty on construction
CORE-1164 introduced an optimization where initialization
of CBlockIndex doesn't mark it as dirty since it might
never be used. In reality there are cases where descendant
is stored into the database and in such case after restart
we'd be missing a block index thus forming a broken chain
which is not supported and results in an assertion termination.