Fix duplicate transaction notification.
We were recording the ID of a transaction we had sent a notification for
too early, this meant that if validation of a double-spend failed we would
not attempt a notification for a later valid double-spend because we
erroneously thought we had already performed a notification.

Don't record a double-spend notification attempt until we actually make
the attempt.