Remove -allowunsolicitedaddr config option.
The -allowunsolicitedaddr config option was introduced to force a fallback
to the old behaviour in the event of receiving an unsolicited ADDR message
from a peer. It was only a temporary measure, and after 2 years without issue
it should be safe to remove and make rejecting such messages the standard
behaviour.