Implement throttling within the JBA when it exceeds a certain threshold.
When the JBA exceeds the configured threshold it will start throttling the
rate at which transactions are added to the next block template. In
addition, it will ensure that whenever it does add a transaction (or group
of transactions) to the template that transaction will be recent enough
to ensure the block template cannot be considered selfish.

The JBA now supports skipping of transactions (or groups of transactions)
that could be considered selfish. Such transactions will not be included
in the current block template, but will be reconsidered for inclusion in
the next.