Shard the CoinsViewCache.
The cache is made up of a number of independent shards, each of which
can be operated on by a single thread. At the end of a multi-threaded
operation all shards should be folded back into a single cache.

Construct a CoinsViewCache with a single shard by default, and so as to
keep the existing interface available for non-sharding users, forward any
calls directly made to CoinsViewCache to that shard.