Change handling of full send and receive queues.
The old handling of the case where both send and receive queues for a peer
are full was to temporarily increase the size limit on the sending queue to
try and free things up. This approach was always a sub-optimal hack IMHO,
and leaves open the door for a DoS.

Change the handling for the full send and receive queues so that we will
instead continue to pull messages off of the read queue to free it up,
but will drop any further replies to that peer until they start processing
what we have already sent them.